+++ changelog.md
@@ -1,3 +1,11 @@
 # Changelog
 
-* Deleted 'summary.md' file.
+**Test Status**
+* G - Controller Files
+	* 'Connected Device Class' is recognized successfully.
+	* Actual function tests still fail.
+
+---
+
+**./fox-devices/_classes/device.class.js**
+* Wrote basic class structure.